add 激活装备增加等级验证，背包接口 玩家装备新增神谕值展示，增加装备等级展示 update 提升战力系统修改神谕值计算公式

+++ gm-domain/src/main/java/com/gm/modules/basicconfig/entity/EquipmentInfoEntity.java
@@ -45,7 +45,7 @@ public class EquipmentInfoEntity implements Serializable {
 	/**
 	 * 装备等级
 	 */
-	private Long equipLevel;
+	private Integer equipLevel;
 	/**
 	 * 装备图片地址
 	 */
