+++ gm-domain/src/main/java/com/gm/modules/user/rsp/UserEquipInfoDetailRsp.java
@@ -34,7 +34,7 @@ public class UserEquipInfoDetailRsp {
 	/**
      * 装备等级
 	 */
-	private Long equipLevel;
+	private Integer equipLevel;
 	/**
 	 * 装备图片地址
 	 */
